requirements.txt
accelerate>=1.2.1
datasets>=3.1.0
matplotlib>=3.9.2
numpy>=2.1.3
Pillow>=11.0.0
pyyaml>=6.0.2
scikit-learn>=1.5.2
scipy>=1.14.1
torch>=2.5.1
torchvision>=0.20.1
transformers>=4.47.0
peft>=0.13.2
tqdm>=4.67.0
